lift value restriction on kv api
