--- crates/runtime/src/plugins/antiraid/kv/types.rs
@@ -1,31 +1,0 @@
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct KvRecord {
-    pub key: String,
-    pub value: serde_json::Value,
-    pub created_at: Option<chrono::DateTime<chrono::Utc>>,
-    pub last_updated_at: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-impl From<crate::traits::ir::kv::KvRecord> for KvRecord {
-    fn from(record: crate::traits::ir::kv::KvRecord) -> Self {
-        KvRecord {
-            key: record.key,
-            value: record.value,
-            created_at: record.created_at,
-            last_updated_at: record.last_updated_at,
-        }
-    }
-}
-
-impl From<KvRecord> for crate::traits::ir::kv::KvRecord {
-    fn from(record: KvRecord) -> Self {
-        crate::traits::ir::kv::KvRecord {
-            key: record.key,
-            value: record.value,
-            created_at: record.created_at,
-            last_updated_at: record.last_updated_at,
-        }
-    }
-}
