+++ crates/runtime/src/traits/ir/kv.rs
@@ -1,7 +1,9 @@
+use crate::utils::khronos_value::KhronosValue;
+
 /// Represents a full record complete with metadata
 pub struct KvRecord {
     pub key: String,
-    pub value: serde_json::Value,
+    pub value: KhronosValue,
     pub created_at: Option<chrono::DateTime<chrono::Utc>>,
     pub last_updated_at: Option<chrono::DateTime<chrono::Utc>>,
 }
